requirements.txt
boto==2.49.0
boto3==1.13.11
botocore==1.16.11
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
docutils==0.15.2
Flask==1.1.2
gensim==3.8.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
numpy==1.18.4
Pillow==7.1.2
pytesseract==0.3.4
python-dateutil==2.8.1
pyttsx3==2.87
requests==2.23.0
s3transfer==0.3.3
scipy==1.4.1
six==1.14.0
smart-open==2.0.0
urllib3==1.25.9
Werkzeug==1.0.1
